Add an assert on the value of the log2scale tensor in the FP8 cast ops whenever throwIfLog2ScaleNotInRange is enabled
Summary:
Currently, if the `throwIfLog2ScaleTensorNotInRange` session option is enabled, the FP8 matmul and conv ops will raise a poplar error if the log2 scale value is outside the 6 bit signed integer range.

The FP8 cast ops were missing this so the same check has been added for both `cast_then_pow2scale` and `pow2scale_then_cast`.

!ci_branch_mk2

Test Plan: Tests added

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, matthewha

Maniphest Tasks: T71909

Differential Revision: https://phabricator.sourcevertex.net/D78550

+++ tests/unittests/python/popxl/ops/test_cast.py
@@ -31,7 +31,7 @@ class TestCast:
 
         with g:
             a = popxl.variable(1, popxl.float16, name="t0")
-            log2_scale = popxl.constant(0, popxl.int8, name="log2_scale")
+            log2_scale = popxl.constant(0, popxl.int32, name="log2_scale")
             b = ops.pow2scale_then_cast(a, log2_scale, dtype)
 
         assert b.dtype == dtype
@@ -56,7 +56,7 @@ class TestCast:
             )
             data = host_pow2scale_then_cast(data, from_dtype, 0)
             a = ops.host_load(input0, "a")
-            log2_scale = popxl.constant(0, popxl.int8, name="log2_scale")
+            log2_scale = popxl.constant(0, popxl.int32, name="log2_scale")
             b = ops.cast_then_pow2scale(a, log2_scale, to_dtype)
 
         assert b.dtype == to_dtype
